website/DataService/requirements.txt
astroid==2.15.1
azure-functions==1.13.3
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
dill==0.3.6
dnspython==2.3.0
Flask==2.2.3
idna==3.4
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-object-proxy==1.9.0
MarkupSafe==2.1.2
mccabe==0.7.0
platformdirs==3.2.0
pylint==2.17.1
pymongo==4.3.3
requests==2.28.2
tomli==2.0.1
tomlkit==0.11.7
typing_extensions==4.5.0
urllib3==1.26.15
Werkzeug==2.2.3
wrapt==1.15.0
